+++ components/app/result/dish-comment-add-section.tsx
@@ -1,9 +1,12 @@
 import { useLocalSearchParams, useRouter } from 'expo-router';
 import { useEffect, useState } from 'react';
-import { StyleSheet, Text } from 'react-native';
+import { Alert, StyleSheet, TouchableOpacity, View } from 'react-native';
+import { Input } from '@/components/common/input';
+import { Loader } from '@/components/common/loader';
+import { Icon } from '@/components/common/icon';
+import { postNewComment } from '@/api/api';
 
 interface DishCommentAddSectionProps {
-    // refetch: () => Promise<QueryObserverResult<MealComment[], Error>>;
     refetch: () => void;
 }
 
@@ -20,47 +23,57 @@ export function DishCommentAddSection({ refetch }: DishCommentAddSectionProps) {
         }
     }, [isPosting]);
 
-    // FIXME: Remove this
-    const onKeyDown = (e: KeyboardEvent): void => {
-        // if (e.key === 'Enter') {
-        //     onAddComment();
-        // }
-    };
-
     const onAddComment = async () => {
         setIsPosting(true);
 
         try {
-            // await postNewComment({
-            //     mealId: id,
-            //     text: commentValue
-            // });
-            //
-            // await refetch();
+            await postNewComment({
+                dishId: id,
+                text: commentValue
+            });
+
+            refetch();
         } catch (err: unknown) {
             // if (err instanceof ApiError) {
             //     handleApiError(err, router, userContext);
             // }
+            // Alert.alert(err.message);
+            // NOTE: Possibly tokens do not match
+            // FIXME: Refresh token every some time
+            Alert.alert('Błąd. Komentarz nie został dodany.');
         } finally {
             setIsPosting(false);
         }
     };
 
     return (
-        // <div className={styles['meal-comment-add-section']}>
-        //     <div className={styles['meal-comment-add-section__input']}>
-        //         <InputString label={'Nowy komentarz'} value={commentValue} setValue={setCommentValue} onKeyDown={onKeyDown} />
-        //         <div className={styles['meal-comment-add-section__send-icon']}>
-        //             {isPosting ? <Loader /> : <SendIconButton onClick={onAddComment} disabled={commentValue.length === 0} />}
-        //         </div>
-        //     </div>
-        // </div>
-        <Text>X</Text>
+        <View style={styles['dish-comment-add-section']}>
+            <View style={styles['dish-comment-add-section__input']}>
+                <Input label={'Nowy komentarz'} value={commentValue} onChange={setCommentValue} />
+                <View style={styles['dish-comment-add-section__send-icon']}>
+                    {isPosting
+                        ? <Loader size={30} />
+                        : (
+                            <TouchableOpacity onPress={onAddComment} disabled={commentValue.length === 0}>
+                                <Icon type="post" size={32} />
+                            </TouchableOpacity>
+                        )
+                    }
+                </View>
+            </View>
+        </View>
     );
 }
 
 const styles = StyleSheet.create({
-    'meal-comment-add-section': {},
-    'meal-comment-add-section__input': {},
-    'meal-comment-add-section__send-icon': {}
+    'dish-comment-add-section': {
+        marginVertical: 8
+    },
+    'dish-comment-add-section__input': {
+        flexDirection: 'row'
+    },
+    'dish-comment-add-section__send-icon': {
+        marginTop: 24,
+        marginLeft: 8
+    }
 });