+++ components/app/result/dish-comment-container.tsx
@@ -1,18 +1,16 @@
-import { useLocalSearchParams } from 'expo-router';
-import { Dimensions, StyleSheet, View } from 'react-native';
-import { DishComment } from '@/types/dish.types';
+import { Link, useLocalSearchParams } from 'expo-router';
+import { Dimensions, StyleSheet, Text, View } from 'react-native';
 import { Heading } from '@/components/common/heading';
 import { Loader } from '@/components/common/loader';
 import { DishCommentContent } from '@/components/app/result/dish-comment-content';
+import { isLoggedIn } from '@/contexts/user.context';
+import { DishCommentAddSection } from '@/components/app/result/dish-comment-add-section';
+import { spanishBlue } from '@/constants/colors';
+import { useGetDishComments } from '@/hooks/use-get-dish-comments';
 
 export function DishCommentContainer() {
     const { id } = useLocalSearchParams<{ id: string }>();
-    // const { isLoggedIn } = useUserContext();
-    const isLoggedIn = () => false;
-    // const { data, isLoading, refetch } = useGetMealComments(id);
-    const data: DishComment[] = [];
-    const isLoading = false;
-    const refetch = () => {};
+    const { data, isLoading, refetch } = useGetDishComments(id);
 
     return (
         <View style={styles['dish-comment-container']}>
@@ -20,10 +18,10 @@ export function DishCommentContainer() {
                 ? (
                     <View style={styles['dish-comment-content']}>
                         <Heading level={3} style={styles['dish-comment-header']}>Komentarze</Heading>
-                        {/*{isLoggedIn()*/}
-                        {/*    ? <MealCommentAddSection refetch={refetch} />*/}
-                        {/*    : <Text style={{ textAlign: 'center' }}><Link href={'/users/login'}>Zaloguj się</Link>, aby dodać komentarz.</Text>*/}
-                        {/*}*/}
+                        {isLoggedIn()
+                            ? <DishCommentAddSection refetch={refetch} />
+                            : <Text style={styles['text-about-login']}><Link href={'/login'} style={styles['link-to-login']}>Zaloguj się</Link>, aby dodać komentarz.</Text>
+                        }
                         <DishCommentContent data={data} />
                     </View>
                 )
@@ -42,5 +40,13 @@ const styles = StyleSheet.create({
     'dish-comment-content': {},
     'dish-comment-header': {
         textAlign: 'center'
+    },
+    'text-about-login': {
+        textAlign: 'center',
+        marginBottom: 16
+    },
+    'link-to-login': {
+        color: spanishBlue,
+        textDecorationLine: 'underline'
     }
 });